requirements.txt
streamlit
langchain
langchain-xai
langchain-huggingface
faiss-cpu
langchain-community     # Required for TextLoader, PyPDFLoader, etc.



aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
appnope==0.1.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.3.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.8
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.5
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.2
executing==2.0.1
faiss-cpu==1.9.0
filelock==3.15.4
Flask==3.1.0
frozenlist==1.5.0
fsspec==2024.6.1
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.0
idna==3.8
importlib_metadata==8.4.0
ipykernel==6.29.5
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.2
jupyter_core==5.7.2
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.28
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.4
langgraph==0.2.60
langgraph-checkpoint==2.0.9
langgraph-sdk==0.1.48
langsmith==0.2.6
MarkupSafe==2.1.5
marshmallow==3.23.2
matplotlib-inline==0.1.7
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
openai==1.58.1
orjson==3.10.12
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
Pygments==2.18.0
pyparsing==3.2.0
pypdf==5.1.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.7.24
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.13.1
sentence-transformers==3.3.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
stack-data==0.6.3
sympy==1.13.2
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.21.0
torch==2.4.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.1
typing-inspect==0.9.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==3.1.3
yarl==1.18.3
zipp==3.20.0
gunicorn==23.0.0




python-dotenv 
langchain 
langchain-community 
langchain-huggingface 
langchain-groq


flask
PyPDFLoader
 langchain_google_genai
langchain-community
sentence-transformers
faiss-cpu
langchain
langchainhub


pip install langchain-groq==0.1.5
pip install langchain==0.2.16
pip install langchain-community==0.2.16
pip install langchain-core==0.2.43
pip install langchain-text-splitters==0.2.4

pip install langchain
pip install langchain-core
pip install langchain-community
pip install langchain-huggingface
pip install faiss-cpu
pip install pypdf
pip install transformers
pip install flask
pip install sentence-transformers
pip install langchain_groq